% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sisti-package.R
\docType{package}
\name{sisti-package}
\alias{sisti}
\alias{sisti-package}
\title{sisti: Real-Time PCR Data Sets by Sisti et al. (2010)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Other contributors:
\itemize{
  \item Pattern Institute [copyright holder, funder]
}

}
\keyword{internal}
